Bug fixing
Changes in this commit
-------------------------------------------------------------------------------

    - Fixed issue with the blobRead function template in core/include/fileops.h. The function now
        takes a reference to a pointer to a char as an argument for the memory buffer. Doing this
        allows changes made to the buffer address to persist into the calling scope. This issue was
        causing texture loading to fail as a secondary effect.

    - Fixed Phong shading program, it now compiles.

    - RenderNodes, as well as most other rendering objects now hold references in std::shared_ptr's
        instead of plain references. This was done to counteract an issue where certain objects
        were being deleted more than once.

    - Changes made to Effect and derived classes to make binding effects more straightforward.
        Effects must now be explicitly bound before use, instead of being implicitly bound during
        construction.

    - There is now a distinction between uniform blocks/buffers that need to be updated every frame
        versus once per object. The RenderSystem exposes a vector which stores shared pointers to
        BlockDefinitions in order to accomodate the new distinction.

Known Issues
-------------------------------------------------------------------------------

    - Nothing appears on the screen. I'm not sure why, there should be a cube. Or something.

+++ rendering/src/texture.cpp
@@ -84,7 +84,7 @@ void Texture::loadTGA(char *buffer)
         return;
     }
     cout << "Image Properties: Origin (" << header.x_origin << " ," << header.y_origin << "), "
-         << header.width << "x" << header.height << " " << header.depth << " bits per pixel, " 
+         << header.width << "x" << header.height << " " << (int)header.depth << " bits per pixel, " 
          << (header.descriptor & 56) << " bits alpha." << endl;
 
     // Skip over the image ID if present, then push data to GPU
